PIM-10897: Move services to Api/query folder
